# Conflicts:
#	src/main/java/tad/controller/UserGuestController.java

+++ src/main/java/tad/controller/UserCartController.java
@@ -11,7 +11,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import tad.DAO.IAddressDAO;
 import tad.DAO.ICartDAO;
 import tad.entity.Account;
 import tad.entity.Cart;
@@ -24,15 +23,14 @@ public class UserCartController {
 	@Autowired
 	private ICartDAO cartDAO;
 
-	@Autowired
-	private IAddressDAO addressDAO;
-
 	@RequestMapping(value = "index")
 	public String cart(ModelMap modelMap, HttpSession session) {
 
 		float subtotal = 0;
 		Account account = (Account) session.getAttribute("account");
-
+		if (account == null) {
+			return "redirect:/admin/overview.htm";
+		}
 		List<Cart> list = cartDAO.getCart(account.getAccountId());
 
 		int canCheckOut = 1;
@@ -41,7 +39,11 @@ public class UserCartController {
 			for (Cart c : list) {
 				// Chi tinh nhung sp con hang
 				if (c.getProduct().getQuantity() > 0) {
-					subtotal += c.getQuantity() * c.getProduct().getPrice();
+					if (c.getProduct().getCoupon() != null && c.getProduct().getCoupon().checkVaildCoupon())
+						subtotal += c.getQuantity() * (c.getProduct().getPrice()
+								- (c.getProduct().getPrice() * c.getProduct().getCoupon().getDiscount()));
+					else if (c.getProduct().getCoupon() == null || (!c.getProduct().getCoupon().checkVaildCoupon()))
+						subtotal += c.getQuantity() * c.getProduct().getPrice();
 
 				}
 
@@ -67,22 +69,31 @@ public class UserCartController {
 	@RequestMapping(value = "delete/{productID}.htm")
 	public String delete(ModelMap model, HttpSession session, @PathVariable("productID") String productID) {
 		Account account = (Account) session.getAttribute("account");
+		if (account == null) {
+			return "redirect:/admin/overview.htm";
+		}
 		cartDAO.deleteCart(cartDAO.getCart(account.getAccountId(), Integer.parseInt(productID)));
 		List<Cart> list = cartDAO.getCart(account.getAccountId());
 		model.addAttribute("carts", list);
+		int totalCart = (int) session.getAttribute("totalCart");
+		if (totalCart > 0)
+			session.setAttribute("totalCart", totalCart - 1);
 		return "redirect:/cart/index.htm";
 	}
 
 	@RequestMapping(value = "quantity/plus", params = { "productID", "quantity" })
-	public String qtyPlus(ModelMap model, @RequestParam("productID") int id, @RequestParam("quantity") int qty) {
+	public String qtyPlus(ModelMap model, HttpSession session, @RequestParam("productID") int id,
+			@RequestParam("quantity") int qty) {
 		cartDAO.updateQuantity(id, qty + 1);
 
 		return "redirect:/cart/index.htm";
 	}
 
 	@RequestMapping(value = "quantity/minus", params = { "productID", "quantity" })
-	public String qtyMinus(ModelMap model, @RequestParam("productID") int id, @RequestParam("quantity") int qty) {
+	public String qtyMinus(ModelMap model, @RequestParam("productID") int id, HttpSession session,
+			@RequestParam("quantity") int qty) {
 		cartDAO.updateQuantity(id, qty - 1);
+
 		return "redirect:/cart/index.htm";
 	}
